Rename vector-store module to experimental_vector-store
