+++ packages/giselle-engine/src/core/experimental_vector-store/paths.ts
