+++ packages/giselle-engine/src/core/experimental_vector-store/types/interface.ts
